Build efi.bin for systemd-boot platforms
Some platforms such as AIM series use efi.bin instead of boot.img.
So efi.bin building function should be added to this tool.
Esdk is essential in building efi.bin, so it's extraction and installation
process is involved in this patch. It should be noted that some tools may
evolve over time, however the version of esdk was fixed in this patch and
this is a workaround. It should be optimized later.

Signed-off-by: Yijie Yang <quic_yijiyang@quicinc.com>